+++ tests/ui/redundant_static_lifetimes.rs
